requirements.txt
annotated-types==0.7.0
anyio==4.8.0
blinker==1.9.0
Bootstrap-Flask==2.4.1
certifi==2024.12.14
click==8.1.8
colorama==0.4.6
distro==1.9.0
Flask==3.1.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-WTF==1.2.2
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
MarkupSafe==3.0.2
openai==0.28.0
pydantic_core==2.27.2
pydantic==2.10.5
python-dotenv==1.0.1
requests==2.31.0
sniffio==1.3.1
SQLAlchemy==2.0.37
tqdm==4.67.1
typing_extensions==4.12.2
Werkzeug==3.1.3
WTForms==3.2.1
